third_party/WebKit/LayoutTests/external/wpt/fullscreen/api/element-request-fullscreen-not-allowed-expected.txt

This is a testharness.js-based test.
FAIL Element#requestFullscreen() when not allowed to request fullscreen assert_equals: event.target expected Element node <div id="log"></div> but got Document node with 2 children
Harness: the test ran to completion.
